nmap_script_puzzying/requirement.txt
appdirs==1.4.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
dnspython==1.16.0
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
jdcal==1.4.1
openpyxl==3.0.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
virtualenv==20.0.7
zipp==3.0.0